requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-community
langchain-openai
python-dotenv
openai
pydantic
sentence-transformers
langchain-mistralai

langchain-huggingface